requirements.txt
Django==5.1.2
django-filter==24.3
djangorestframework==3.15.2
drf-spectacular==0.27.1
psycopg==3.2.3
python-decouple==3.8
ruff==0.5.3